% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartment_2.R
\name{graph_2c}
\alias{graph_2c}
\title{Regression model of the pharmacokinetics
Use the model of 2 compartment}
\usage{
graph_2c(data, time, conc)
}
\arguments{
\item{data}{data frame, which include time and concentrations points}

\item{time}{time row from data frame}

\item{conc}{concentration row from data frame}
}
\value{
an exponential regression model of experimental data
}
\description{
Regression model of the pharmacokinetics
Use the model of 2 compartment
}
\examples{
\dontrun{
long_running_function()
}
}
