% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_3c.R
\name{graph_3c}
\alias{graph_3c}
\title{Regression model of the pharmacokinetics}
\usage{
graph_3c(data, time, conc)
}
\arguments{
\item{data}{data frame, which include time and concentrations points}

\item{time}{time row from data frame}

\item{conc}{concentration row from data frame}
}
\value{
an exponential regression model of experimental data
}
\description{
Use the model of 3 compartment
}
\examples{
\dontrun{
long_running_function()
}
}
