% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartment_1.R
\name{par_1c}
\alias{par_1c}
\title{Parameters of the pharmacokinetics}
\usage{
par_1c(data, time, conc)
}
\arguments{
\item{data}{data frame, which include time and concentrations points}

\item{time}{time row from data frame}

\item{conc}{concentration row from data frame}
}
\value{
PK parameters%: A, alpha, R^2 of the regression line, AUC, CL1, V1, K10, T1/2, Cmax
}
\description{
Use the model of 1 compartment
}
\examples{
\dontrun{
long_running_function()
}
}
