% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartment_2.R
\name{par_2c}
\alias{par_2c}
\title{Parameters of the pharmacokinetics}
\usage{
par_2c(data, time, conc)
}
\arguments{
\item{data}{data frame, which include time and concentrations points}

\item{time}{time row from data frame}

\item{conc}{concentration row from data frame}
}
\value{
PK parameters%: "A", "alpha", "B", "beta", "R^2", "AUC", "K21", "K10", 
"K12", "V1", "V2", "Vdss", "CL1", "CL2", "T1/2 alpha", "T1/2 beta"
}
\description{
Use the model of 2 compartment
}
\examples{
\dontrun{
long_running_function()
}
}
